Refactor: repository package path and delete duplicated
