--- src/main/java/com/lckback/lckforall/aboutlck/repository/PlayerRepository.java
@@ -1,8 +1,0 @@
-package com.lckback.lckforall.aboutlck.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.lckback.lckforall.player.model.Player;
-
-public interface PlayerRepository extends JpaRepository<Player, Long> {
-}
