+++ src/main/java/com/lckback/lckforall/aboutlck/service/AboutLckTeamService.java
@@ -15,9 +15,9 @@ import com.lckback.lckforall.aboutlck.dto.team.FindTeamPlayerHistoryDto;
 import com.lckback.lckforall.aboutlck.dto.team.FindTeamRatingHistoryDto;
 import com.lckback.lckforall.aboutlck.dto.team.FindTeamRatingBySeasonDto;
 import com.lckback.lckforall.aboutlck.dto.team.FindTeamWinningHistoryDto;
-import com.lckback.lckforall.aboutlck.repository.SeasonRepository;
-import com.lckback.lckforall.aboutlck.repository.SeasonTeamRepository;
-import com.lckback.lckforall.aboutlck.repository.TeamRepository;
+import com.lckback.lckforall.team.repository.SeasonRepository;
+import com.lckback.lckforall.team.repository.SeasonTeamRepository;
+import com.lckback.lckforall.team.repository.TeamRepository;
 import com.lckback.lckforall.base.api.error.CommonErrorCode;
 import com.lckback.lckforall.base.api.exception.RestApiException;
 import com.lckback.lckforall.team.model.Season;
