+++ src/main/java/com/lckback/lckforall/match/service/PogService.java
@@ -10,7 +10,7 @@ import com.lckback.lckforall.match.model.Match;
 import com.lckback.lckforall.match.model.Set;
 import com.lckback.lckforall.match.repository.MatchRepository;
 import com.lckback.lckforall.player.model.Player;
-import com.lckback.lckforall.player.model.repository.PlayerRepository;
+import com.lckback.lckforall.player.repository.PlayerRepository;
 import com.lckback.lckforall.vote.model.MatchPogVote;
 import com.lckback.lckforall.vote.model.SetPogVote;
