+++ src/main/java/com/lckback/lckforall/player/repository/PlayerRepository.java
@@ -1,6 +1,5 @@
-package com.lckback.lckforall.player.model.repository;
+package com.lckback.lckforall.player.repository;
 
-import com.lckback.lckforall.match.model.Match;
 import com.lckback.lckforall.player.model.Player;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
