+++ src/main/java/com/lckback/lckforall/team/repository/TeamRepository.java
@@ -1,4 +1,4 @@
-package com.lckback.lckforall.aboutlck.repository;
+package com.lckback.lckforall.team.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
